+++ routes/web.php
@@ -86,4 +86,4 @@ Route::get('admin/order/list', [App\Http\Controllers\OrderController::class, 'li
 Route::get('admin/item/list/{id}', [App\Http\Controllers\ItemController::class, 'list'])->name('admin.item.list');
 
 // Consuming APIs
-Route::get('consumeApi', [App\Http\Controllers\ThisController::class, 'apiNoKey'])->name('consume.api');
+Route::get('consumeApi', [App\Http\Controllers\Api\ThisController::class, 'apiNoKey'])->name('consume.api');
